Make the DataFormBuildArea component leaner, move styles to DataFormCanvas
