+++ mathesar_ui/src/systems/data-forms/form-maker/DataFormCanvas.svelte
@@ -1,5 +1,8 @@
 <script lang="ts">
-  import { dataFormInspectorWidth } from '@mathesar/stores/localStorage';
+  import {
+    dataFormInspectorVisible,
+    dataFormInspectorWidth,
+  } from '@mathesar/stores/localStorage';
   import { WithPanel } from '@mathesar-component-library';
 
   import {
@@ -10,6 +13,22 @@
   import DataFormInspector from './inspector/DataFormInspector.svelte';
 
   export let dataFormManager: DataFormManager;
+
+  $: editableDataFormManager =
+    dataFormManager instanceof EditableDataFormManager
+      ? dataFormManager
+      : undefined;
+
+  function handleFormSelection(e: MouseEvent) {
+    if (editableDataFormManager) {
+      const { target } = e;
+      if (target instanceof HTMLElement) {
+        if (!target.closest('[data-form-selectable]')) {
+          editableDataFormManager.resetSelectedElement();
+        }
+      }
+    }
+  }
 </script>
 
 <div class="data-form-canvas">
@@ -17,14 +36,16 @@
     bind:sizePx={$dataFormInspectorWidth}
     minSizePx={200}
     maxSizePx={600}
-    showPanel={dataFormManager instanceof EditableDataFormManager}
+    showPanel={editableDataFormManager && $dataFormInspectorVisible}
   >
     <div class="workarea">
-      <DataFormBuildArea {dataFormManager} />
+      <div class="form-holder" on:click={handleFormSelection}>
+        <DataFormBuildArea {dataFormManager} />
+      </div>
     </div>
     <div class="data-form-inspector-panel" slot="panel">
-      {#if dataFormManager instanceof EditableDataFormManager}
-        <DataFormInspector {dataFormManager} />
+      {#if editableDataFormManager}
+        <DataFormInspector dataFormManager={editableDataFormManager} />
       {/if}
     </div>
   </WithPanel>
@@ -42,6 +63,17 @@
     height: 100%;
     overflow: hidden;
     display: grid;
+    border: 1px solid var(--border-color);
+    border-radius: var(--border-radius-l);
+    background: var(--elevated-background);
+    --data_forms__z-index__field-header: 1;
+    --data_forms__z-index__field-add-dropdown-trigger: 2;
+    --data_forms__label-input-gap: 0.5rem;
+    --data_forms__selectable-element-padding: 1rem;
+
+    .form-holder {
+      overflow: auto;
+    }
   }
 
   .data-form-inspector-panel {
